29 April 2021 17:58

+++ database/migrations/2021_04_28_144308_create_sys_personel_table.php
@@ -22,8 +22,8 @@ class CreateSysPersonelTable extends Migration
             $table->float('min_spd');
             $table->float('max_spd');
             $table->integer('total_personel');
-            $table->float('total_pramu');
-            $table->float('total_kasir');
+            $table->integer('total_pramu');
+            $table->integer('total_kasir');
             $table->timestamps();
         });
     }
